Move translation upload code to util module
