+++ manage.py
@@ -22,8 +22,12 @@ from portal.config.site_persistence import SitePersistence
 from portal.extensions import db, user_manager
 from portal.factories.app import create_app
 from portal.models.clinical_constants import add_static_concepts
-from portal.models.i18n import smartling_download, smartling_upload
-from portal.models.i18n_utils import compile_pos, download_all_translations
+from portal.models.i18n_utils import (
+    compile_pos,
+    download_all_translations,
+    smartling_download,
+    smartling_upload,
+)
 from portal.models.intervention import add_static_interventions
 from portal.models.organization import add_static_organization
 from portal.models.relationship import add_static_relationships
